requirements.txt
newspaper3k==0.2.8
python-dotenv==1.0.0
Requests==2.31.0
streamlit==1.27.0
